django/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
caugetch==0.0.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clipboard==0.0.4
colorama==0.4.6
coverage==7.6.5
Django==4.2.16
django-core==1.4.1
django-cors-headers==4.6.0
django-db-logger==0.1.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-spectacular==0.27.2
exceptiongroup==1.2.2
Faker==12.0.1
fuzzywuzzy==0.18.0
inflection==0.5.1
iniconfig==2.0.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
levenshtein==0.26.1
MarkupSafe==3.0.2
mixer==7.2.2
packaging==24.2
pluggy==1.5.0
prompt-toolkit==3.0.48
psycopg2==2.9.10
PyJWT==2.9.0
pyperclip==1.9.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.10.1
referencing==0.35.1
rpds-py==0.21.0
six==1.16.0
sqlparse==0.5.2
tomli==2.1.0
typing-extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
